Refactor budget context imports and fix income calculation bug
